[Tab Group Sync] Fix crash when unmapping local IDs
The fact a hidden tab group has no local ID was overlooked and we can
crash if a hidden tab group exists and we trigger this behavior.

For context this code triggers if closing the second to last Chrome
instance so we can stop treating all the other tab groups from other
windows as if they don't exist in the Tab Group Pane. This is a
workaround in place until we can actively access multi instance tab
model selectors during runtime.

Fixed: 345711650
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5607826
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1312085}

+++ chrome/browser/tab_group_sync/android/java/src/org/chromium/chrome/browser/tab_group_sync/TabGroupSyncUtils.java
@@ -96,6 +96,9 @@ public final class TabGroupSyncUtils {
 
         for (String syncGroupId : tabGroupSyncService.getAllGroupIds()) {
             SavedTabGroup savedTabGroup = tabGroupSyncService.getGroup(syncGroupId);
+            // If there is no local ID the group is already hidden so this is a no-op.
+            if (savedTabGroup.localId == null) continue;
+
             if (!isInCurrentWindow(filter, savedTabGroup.localId)) {
                 tabGroupSyncService.removeLocalTabGroupMapping(savedTabGroup.localId);
             }
